requirements.txt
torch>=2.0
torchvision
Pillow
matplotlib
pytorch-msssim
compressai
